+++ lib/src/features/generation_seed/core/constants/enums/seed_type_enum.dart
@@ -38,8 +38,8 @@ library;
 import 'package:gui/src/core/utils/gen/localization/locale_keys.dart';
 
 enum SeedTypeEnum {
-  twelve(1, LocaleKeys.seed12words, 12),
-  twentyFour(2, LocaleKeys.seed24words, 24);
+  twelve(1, LocaleKeys.seed_twelve_words, 12),
+  twentyFour(2, LocaleKeys.seed_twenty_four_words, 24);
 
   const SeedTypeEnum(
     this.id,